Restore older debug info fields to their original positions
The declname field had been inserted at the beginning of the debug
information, but now it is appended to the end. This preserves the
order of the existing fields. This is a defensive move to avoid bugs
due to unknowingly disturbing invariants in the debug info structure.
(Lame, I know, but I have other fish to fry right now.)