Detect nodes uniqueness as a function of the schema
A helper class Marker is used to mark whether nodes are unique or
not. When 'uses' nodes are expanded, if they're repeated then their
hanging leaves nodes are marked as non unique.

Also created method `print_entry` that is used to print scalar nodes
and array nodes.  In the case of array nodes, position is prepend
with the full node path is node is not unique.