Simplify alarm notifications listener
Instead of having to RPC and then turn the socket into a listening
state, we should just have a separate socket for notifications.  It's
stateless anyway.  No need for the special length-prefixed message
protocol either -- we can just write JSON and trust fibers to allow
concurrency.  The listen program likewise simplifies to a glorified
socat.