Determine whether a configuration should run on-a-stick
A configuration should run on-a-stick always is on-a-stick command line
argument is uses, regardless of device setup in configuration.  The
arguments can only be used if the configuration has one instance.

In case on-a-stick argument is not used, a lwAFTR runs on-a-stick if
internal and external interfaces use the same device.

In case of running on-a-stick:
- Always use a V4V6 splitter app to separater IPv4 and IPv6 traffic.
- Except if internal and external interfaces use VLAN tags and the VLAN
  tags are different.