Initialize stats queue number
Arguments rxcounter/txcounter indicate which stats queue is used to
store NIC statistics such as rxpacket, txpackets, etc.  If these
arguments are not initialized all the stats are stored in queue 0.
When using RSS which uses a NIC with the same PCI address, if this
argument is not initialized all virtualized NICs use queue 0 for stats,
making it impossible to differentiate statistic per queue.  Thus, these
arguments are initialized to queue id.