Modify generate-binding-table to generate a configuration file instead
Extends generate-binding-table command to allow generation of lwAFTR
configuration file, thus it's not necessary to heavily ammend the
generated binding table, which can be very troublesome if the file
is large.

The old behaviour of generate-binding-table is preserved though, mainly
for backwards compatibility.  Perhaps should be removed in the future.