Restrict condition for use of V4V6 splitter
So far a V4V6 splitter was needed if VLAN tags for external and
internal interfaces were equal.  This condition assumed an scenario
where MAC addresses for external and internal interfaces were equal too.
It is necessary to add this constraint to the condition that evaluates
whether a splitter is necessary.

Summarizing, we are only going to use a V4V6 splitter when both VLAN
tags and MAC addresses of external and internal interfaces are the same.