snabb top: refactor queue stats filtering
Use a simpler method to figure out which queue stats to filter
out. Revert change to symlink each counter for intel_mp and
just symlink the whole folder again. Then add a counter in the
per-app shm that just stores the rxcounter/txcounter.

Snabb top then looks through all the apps in a process and filters
out queue counters that don't match any app. This is not totally precise
in the case that a process uses multiple NICs with counters set
(it may show more counters than necessary).