connectx4.lua: fix page sizes in CREATE_{C,R}Q
Set the page size parameters for CQs and RQs so that the entire queue
fits in a single page.  The alternative would be to select a fixed
page size and add as many PAS entries as necessary.  It is unclear
whether there is a difference between the methods (e.g. for
perfomance).