intel_mp app instantiations make stats symlink counter forest
This commit changes the intel_mp driver to make a forest of symlinks to
the shared stats counters, rather than a single link to the shared
counters directory.  This allows the app to indicate which queue it's
on, as well as allowing ptree and top to naturally track the set of
counters used by the app so they can manage associated .rrd files.

As part of this change, we remove the instance-local "pci" directory,
instead using per-app "pci" directories.  This allows us to more nicely
record which app is on which device.

This commit also updates "snabb top" to get PCI interfaces from within
apps' SHM frames.