Show nexthop MAC in top
It's desirable to be able to retrieve nexthop information via `snabb
config get-state` too. According to the state information of the schema,
information about queues, interfaces and resolved-macs should also be
printed. However this information is not collected yet.

I tried to append at least the resolved mac but I couldn't figure out
how to do it. Thus, I marked the missing feature as TODO.