program.l2vpn: major rewrite, basic IPv4 support
The pseudowire app is replaced by a more generic de-multiplexing
architecture.  The main goals are to make the system more modular and
make it more resilient to performance drops due to shifting workloads.

Packet headers are now removed in two stages (IP header/encapsulation
header) and the pseudowire control-channel is moved to a separate app.

The system is now completelt dual-stack apart from
fragmentation/reassembly for IPv4.