apps.ipv4.fragment: improve performance, reduce garbage
The main processing loop is split into three separate loops to reduce
the effect of unbiased branches due to workloads with either mixed
fragmented/unfragmented packets or a shift from mainly fragmented to
mainly unfragmented packets on the selection of traces.

All packets needing fragmentation are processed in a separate
loop. The fragmentation itself contains another loop which might get
compiled first.  In that case, the pointers passed to
fragment_and_transmit() would require dynamic boxing on the Lua stack,
which cause GC activity.  Explicit boxing is used to avoid this.