luajit: disable function compilation
Compiled functions have almost no benefit in an application which is
essentially 100% compiled. This is because they are only called from
the interpreter and never from other compiled code (function calls are
always inlined by the trace recorder). But they can have an adverse
effect on the process of finding other traces because they force the
system to spend more time trying to compile traces that cannot
succeed.