core.app: replace the main loops with "goto"
The loops that should be compiled most efficiently are those that
process packets in the pull() and push() methods of apps.  The app
engine wraps two levels of loops around them, the breathe() loop and
the pull_order and push_order loops, respectively. Letting the system
figure out that the outer loops cannot be compiled as looping traces
strains the compiler and causes a significant number of aborts in
trace recordings.  By replacing those outer loops with "straight" code
paths connected with "goto", we relieve the compiler from that task
and let it find the inner loops more easily.