core.app: reinstate main breathe loop
The mechanism introduced in 06b5c782 replaces the push/pull loops as
well as the main loop with gotos.  Further analysis has shown that
there is no benefit in flattening the main loop in this manner, in
fact it can even have a negative impact on loop compilation.  This
commit reinstatest the original repeat/until breathe loop.