lib.yang.schema: implement semantics for enumeration type
Implement semantics according to RFC6020, 9.6.  When validating a
configuration against a schema which contains

leaf foo {
  type enumeration {
    enum bar;
    enum baz;
  }
}

the resulting Lua table will contain the key "foo" whose value is the
string of the chosen enum in the configuration, e.g.

{
  foo = "baz"
}

It is debatable whether this should actually result in the numerical
value assigned to the name in the schema.  The name appears to be more
useful because the consumer of this table would have to reproduce the
same name-to-value mapping as the schema, which introduces a source of
possible errors.