program.l2vpn: pass src/dst to tunnel unknown packet handler
This the identification of the tunnel endpoints from the error
messages.  With the new scheme to use a dedicated VC ID for the
control channel, it is no longer possible to identify a VC ID mismatch
directly.

This also makes transporting the VC ID in the control channel
obsolete, but it is left there to be compatible with older versions.