lib.poptrie: avoid cache thrashing in micro benchmark
The micro benchmark would (with increased numbers of entries) mostly be
cycle-bound by loading keys into cache. In real world scenarios it is probably
more reasonable to excpect keys/addresses to be ready in cache. This change
reduces the number keys to be queried to 100 in order to avoid measuring a cold
cache.

With this change the micro benchmark is possibly overly optimistic since the
entropy in which portions of the poptrie accesses is very low, so it is not
universally representative either (that is how it goes with micro benchmarks.)
On the upside, this way the micro benchmark should be a better indicator for
the effects of changes to the code.