core.app: use implict pull/push loops in breathe()
The goal of this change is to simplify the loop structure of the Snabb
engine to make it more likely that the compiler can find and optimize
the packet-processing loops. In particular, it can help in the
situation when the program starts in an idle state with no packets to
process. In that case, we want to avoid that traces compiled for this
initial workload have a negative impact on the creation of traces for
the actual (non-idle) workload.

See https://github.com/snabbco/snabb/issues/1414 for an analysis with
a simplified engine.