Remove bogus lwaftr benchmarking file
The file in question is too small for the minimum ethernet frame size;
with the CRC it becomes 58 bytes, which is smaller than the minimum
frame size of 64, so it gets padded as it goes out on the wire, which
later gets rejected on the receiver as its lengths don't make sense.  In
any case it doesn't test what it means to test, so remove it.