Detect irqbalance presence and suggest remedy
This patch will print a warning if a network function tries to bind
to a CPU, but irqbalance is detected.  Irqbalance, installed by default
on Debian-derived systems, will modify CPU affinities for IRQ
handlers in an effort to spread out the IRQ-processing load among
CPUs.  However this is not what we want in Snabb; we do not want
data-plane CPUs to run IRQ handlers, and we do not want bug reports
coming from users that have this daemon installed.

See https://github.com/snabbco/snabb/blob/master/src/doc/performance-tuning.md#avoid-interrupts-on-data-plane-cpus, for more details.