core.app: only call push() if the link is not empty
This helps keeping the JIT traces small when apps have nothing to
process.

Many apps also perfom housekeeping functions in push(), which need to
be executed even if no packets are being processed.  Those need to be
moved in a separate method "housekeeping()", which is called
explicitly from the breathe loop.