apps.interlink: conform to lib.ptree semantics
Respect the restrictions on the module's name space imposed
by lib.ptree.action_codec.find_public_name().

The ptree mechanism also transforms a nil argument to an app's new()
method in the master process to an empty table in the worker process.
This commit fixes this by an explicit type check.