program.l2vpn: modularize tunnel protocols
Move all of the specific tunnel protocol code out of l2vpn.lua to
simplify adding new protocols in the future.  To add a new protocol,
it only needs to be added as a new type to the schema.  The code for
it must reside in program/l2vpn/tunnels in a file named <type>.lua.