program.l2vpn: support NAT with IPsec
l2vpn is transparent for NAT if IPsec is not used. With IPsec enabled
and the remote endpoint behind a NAT, the interface with the IKE
daemon must be aware of the inside address of the remote endpoint. A
new element in the endpoint container is added for this purpose.