lwaftr: map ietf instance name to (singleton) lwaftr name
Instead of mapping the bind-instance name to the device, we map it to the name
of the lwaftr instances, effectively mapping the snabb-softwire-v2 to a
ietf-softwire-br config with a single instance.

Removed the ability to remove devices via the ietf-softwire-br schema (no
wholly invisible from the ietf schema), and pre_update translations (for now).

Be more strict about path mappings (match the actual path components, and
assert matching queries).