lib.yang.data: reorder translated members before printing
This fixes a bug in which data_printer_from_grammar would not print choice case
members when an order is given to body_printer.

XXX: it is not clear to me why some invocations of body_printer take orderings,
since it seems that the pre-ordering always matches the internal ordering logic
of body_printer. \o/