apps.mellanox.connectx4: fix mcast counters
The intel_mp driver conflates multicast packets with broadcast
packets, which then needs to be reversed by lib.ipc.shmem.iftable_mib.
We have to do the same here to get conistent MIB counters.  It would
probably be better to properly count the multicast packets in the
Intel driver in the first place.