Add scan-protection feature
If enabled, aggregate number of flows, packets and octets by a given
prefix-length.  If an aggregate exceeds a threshold rate of flows with
a maximum number of packets per flow and bytes per packets, limit the
rate of exported flows from all addresses in the aggregate.

This signature uniquely identifies flows due to scanning activity and
protects the flow collector from being overwhelmed by a high flow rate
(rates of up to 150kps from a single address have been observed in
practice).