Don't export empty flows
After an active timeout of a flow, the flow remains in the table with
byte and packet counts set to zero. This can lead to the exporting of
an empty flow if no new packets arrive until the idle timeout.  This
is avoided by checkingg for a non-zeroo packet count before exporting.