apps.mellanox: Use consumer counter in cxq to simplify ownership handling
Rename cxq.next_rx_cqeid to cxq.rx_cqcc (completion queue consumer
counter) and expand to 32 bit. This counter is not wrapped around the
size of the cq. It is used to calculate the value of the SW ownership
bit directly with efficient bitops and eliminates the need for the
rx_mine member of cxq.

# Conflicts:
#	src/apps/mellanox/connectx.lua