apps.ipfix.maps: re-create map ctable if source file has changed
If mk_map() is called via ipfix:reconfig() because the name of a map
file has changed, the map ctable will still not be re-created due to
its presence in the cache used by mk_map(). This commit uses the ctime
of the file to detect whether the ctable needs to be re-created.