ipfix: graceful reload of maps
Reconfiguration of a ipfix app is currently implemented by re-creating
all data structures. This causes disruption of flow collection even
when it is not strictly necessary. As a first step towards a more
graceful reconfiguration, this commit addresses the situation when the
only change is in the set of mapping files of a template. In that
case, only the mapping tables are reconstructed.