ipfix: add HTTPS templates to extract the TLS SNI
Due to lack of a standard information element, this template re-uses a
private element under the PEN of Progress/Flowmon (39499) to be
compatible with their commercial Netflow collector.

The templates parse the TLS handshake record of the client and look
for the SNI extension in the Client Hello message for TLS 1.2 and
later.