apps.mellanox.connectx: add support for PCAM register
The ethernet port speed is read from the eth_proto_oper field of the
PTYS register. This field was extended with the ext_proto_oper field
in newer devices. In particular, 200G and 400G speeds are contained in
this field rather than eth_ptoto_oper.

Which of these fields is valid for the device is determined by the
field ptys_extended_ethernet provided by the PCAM (Ports CApabilities
Mask) register. Whether this register exists or not is, in turn,
determined by the pcam_reg flag of the HCA general cpabilities.